add create talk
